Refactor: swissknife main() in dedicated compilation unit
